+++ integrations/github/tests/github/actions/test_dispatch_workflow_executor.py
@@ -84,7 +84,7 @@ def executor(
     mock_rest_client: MagicMock,
 ) -> Generator[DispatchWorkflowExecutor, None, None]:
     with patch(
-        "github.actions.dispatch_workflow_executor.create_github_client_for_org",
+        "github.actions.abstract_github_executor.create_github_client_for_org",
         new=AsyncMock(return_value=mock_rest_client),
     ):
         yield DispatchWorkflowExecutor()
@@ -101,7 +101,7 @@ class TestDispatchWorkflowExecutor:
         )
 
         with patch(
-            "github.actions.dispatch_workflow_executor.create_github_client_for_org",
+            "github.actions.abstract_github_executor.create_github_client_for_org",
             new=AsyncMock(return_value=mock_rest_client),
         ) as create_client:
             executor = DispatchWorkflowExecutor()
@@ -407,7 +407,7 @@ class TestLegacyDispatchWorkflowExecutor:
 
         with (
             patch(
-                "github.actions.dispatch_workflow_executor.create_github_client_for_org",
+                "github.actions.abstract_github_executor.create_github_client_for_org",
                 new=AsyncMock(return_value=mock_rest_client),
             ),
             patch(
@@ -461,7 +461,7 @@ class TestLegacyDispatchWorkflowExecutor:
 
         with (
             patch(
-                "github.actions.dispatch_workflow_executor.create_github_client_for_org",
+                "github.actions.abstract_github_executor.create_github_client_for_org",
                 new=AsyncMock(return_value=mock_rest_client),
             ),
             patch(
